Attempt to write SignalR integration tests
