+++ tests/Lewee.Blazor.Tests.App/Program.cs
@@ -11,4 +11,7 @@ builder.AddProject<Projects.Lewee_Blazor_Tests_Api>(ServiceNames.WebApi)
     .WithReference(signalR)
     .WaitFor(signalR);
 
+// Blazor Server web app (uses local SignalR, no Azure dependency)
+builder.AddProject<Projects.Lewee_Blazor_Tests_Web>(ServiceNames.BlazorServerWeb);
+
 await builder.Build().RunAsync();
