+++ tests/Lewee.Blazor.Tests.Integration/Infrastructure/OrderCreatedAction.cs
@@ -1,0 +1,6 @@
+using Lewee.Blazor.Tests.Contracts;
+using Lewee.Contracts.StateManagement;
+
+namespace Lewee.Blazor.Tests.Integration.Infrastructure;
+
+internal sealed record OrderCreatedAction(PizzaOrder Order, Guid CorrelationId) : IMessageReceivedAction;
