--- tests/Lewee.Blazor.Tests.Integration/MessageToActionMapper.cs
@@ -1,35 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using Lewee.Blazor.Messaging;
-using Lewee.Blazor.Tests.Contracts;
-using Lewee.Contracts.StateManagement;
-using Microsoft.Extensions.Logging;
-
-namespace Lewee.Blazor.Tests.Integration;
-
-[SuppressMessage(
-    "Performance",
-    "CA1812: Avoid uninstantiated internal classes",
-    Justification = "Used via DI")]
-internal sealed class MessageToActionMapper : IMessageToActionMapper
-{
-    private readonly ILogger<MessageToActionMapper> logger;
-
-    public MessageToActionMapper(ILogger<MessageToActionMapper> logger)
-    {
-        this.logger = logger;
-    }
-
-    public IMessageReceivedAction Map(object message, Guid correlationId)
-    {
-        this.logger.LogInformation(
-            "SignalR message received: {MessageType}, CorrelationId: {CorrelationId}",
-            message.GetType().Name,
-            correlationId);
-
-        return message switch
-        {
-            PizzaOrder order => new OrderCreatedAction(order, correlationId),
-            _ => null,
-        };
-    }
-}
