+++ tests/Lewee.Blazor.Tests.Web/Components/Pages/Home.razor
@@ -1,0 +1,5 @@
+@page "/"
+
+<h1>Blazor Server Test App</h1>
+
+<p>This is a test application for Blazor Server SignalR integration.</p>
