+++ tests/Lewee.Blazor.Tests.Web/Components/_Imports.razor
@@ -1,0 +1,2 @@
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
