+++ tests/Lewee.Blazor.Tests.Web/OrderCreatedAction.cs
@@ -1,0 +1,6 @@
+using Lewee.Blazor.Tests.Contracts;
+using Lewee.Contracts.StateManagement;
+
+namespace Lewee.Blazor.Tests.Web;
+
+internal sealed record OrderCreatedAction(PizzaOrder Order, Guid CorrelationId) : IMessageReceivedAction;
